pyproject.toml
[tool.poetry]
name = "gptvalidator"
version = "0.1.0"
description = ""
authors = ["Jakob Serlier <37184788+Jakob-98@users.noreply.github.com>"]
packages = [{include = "gptvalidator"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.7"
pydantic = "^1.10.8"
pytest = "^7.3.1"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
